+++ R/strategy_define.R
@@ -49,20 +49,22 @@ define_strategy <- function(...,
 #' @export
 define_strategy_ <- function(transition, states) {
   
-  if (! get_state_number(states) == get_matrix_order(transition)) {
+  if (! get_state_number(states) == length(get_state_names(transition))) {
     stop(sprintf(
-      "Number of state in model input (%i) differ from number of state in transition matrix (%i).",
+      "Number of state in model input (%i) differ from number of state in transition object (%i).",
       get_state_number(states),
-      get_matrix_order(transition)
+      length(get_state_names(transition))
     ))
   }
   
+  
   if (! identical(
-    sort(get_state_names(states)),
-    sort(get_state_names(transition))
+    as.vector(sort(get_state_names(states))),
+    as.vector(sort(get_state_names(transition)))
   )) {
-    stop("State names in model input differ from transition matrix.")
+    stop("State names differ from transition object.")
   }
+  
   
   structure(
     list(
@@ -82,22 +84,22 @@ define_strategy_ <- function(transition, states) {
 #'   object.
 #'   
 #' @keywords internal
-get_matrix <- function(x){
-  UseMethod("get_matrix")
+get_transition <- function(x){
+  UseMethod("get_transition")
 }
 
-get_matrix.default <- function(x){
+get_transition.default <- function(x){
   x$transition
 }
 
-set_matrix <- function(x, m) {
-  UseMethod("set_matrix")
+set_transition <- function(x, m) {
+  UseMethod("set_transition")
 }
 
-set_matrix.default <- function(x, m) {
+set_transition.default <- function(x, m) {
   x$transition <- m
+  x
 }
-
 
 get_states <- function(x){
   UseMethod("get_states")
@@ -113,8 +115,8 @@ set_states <- function(x, s) {
 
 set_states.default <- function(x, s) {
   x$states <- s
+  x
 }
-
 
 get_state_value_names.uneval_model <- function(x) {
   get_state_value_names(get_states(x))