+++ R/transform.R
@@ -1,6 +1,5 @@
 #' Convenience Functions to Compute Probabilities
 #' 
-#' 
 #' These convienience functions make it easier to compute 
 #' transition probabilities from indidence rates, OR, RR, or
 #' probabilities estimated on a different timeframe.
@@ -75,4 +74,33 @@ rr_to_prob <- function(rr, p) {
     res <= 1
   )
   res
+}
+
+#' Rescale Discount Rate
+#' 
+#' Rescale a discount rate between two time frames.
+#' 
+#' Continuous discounting is assumed, i.e. when converting a
+#' long-term discount rate into a short-term rate, we assume
+#' that a partial gain from one short term is 
+#' multiplicatively discounted in all following short terms.
+#' At the same time, we assume the short-term rate is
+#' time-invariant.
+#' 
+#' @param x Discount rate to rescale.
+#' @param from Original time period.
+#' @param to Final time period.
+#'   
+#' @return Rate rescaled under the assumption of compound 
+#'   discounting.
+#'   
+#' @export
+#' 
+#' @examples
+#'   ## 1% monthly interest rate to annual
+#'   rescale_discount_rate(0.01, 1, 12)
+#'   ## 3% annual discount rate to (approximately) weekly 
+#'   rescale_discount_rate(0.03, 52, 1)
+rescale_discount_rate <- function(x, from, to) {
+  ((1 + x) ^ (to / from)) -1
 }