+++ tests/testthat/test_model.R
@@ -40,7 +40,7 @@ test_that(
     expect_output(
       str(mod1),
       "List of 2
- $ transition:List of 4",
+ $ transition",
       fixed = TRUE
     )
     expect_output(
@@ -248,7 +248,7 @@ test_that(
       a, 1-a
     )
     
-    res <- heemod:::eval_matrix(mat, par)
+    res <- heemod:::eval_transition(mat, par)
     
     expect_identical(
       round(res[[1]], 2),